% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhyScore_test.R
\name{rhyScore_test}
\alias{rhyScore_test}
\title{Run gene set enrichment tests based on rhythmicity scores}
\usage{
rhyScore_test(data, exp_des, t2g, group = NULL, bootstrap.n = 10000)
}
\arguments{
\item{data}{a matrix of normalized count data.}

\item{exp_des}{a data.frame describing the experimental design with two columns:
group and time.}

\item{t2g}{a data.frame containing terms of gene sets and the corresponding genes}

\item{group}{a vector containing the names of two groups. The first one should
be the reference group and the second one should be the testing group (eg.
c("WT", "KO")).}

\item{bootstrap.n}{the number of time to perform bootstrapping.}
}
\value{
a data.frame containing the testing result of each gene set.
}
\description{
This function enriches highly rhythmic and phase coherent gene sets within one group,
or highly differential rhythmic and phase coherent gene sets between two groups.
}
